app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-adk
google-genai
weaviate-client>=4.0
